pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tesis-ac"
version = "0.1.0"
description = "Pipeline reproducible para modelar crecimiento urbano con Autómatas Celulares + AG"
authors = [
    {name = "Tu Nombre", email = "tu.email@domain.com"},
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.0",
    "rasterio>=1.3.0",
    "geopandas>=0.13.0",
    "shapely>=2.0.0",
    "deap>=1.4.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "pytest>=7.4.0",
    "pillow>=10.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.0",
    "jupyter>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]